+++ NEWS.md
@@ -5,12 +5,13 @@
 - New general purpose `effectsize()` function.
 - Effectsize for differences have CI methods, and return a data frame.
 - Effectsize for ANOVA all have CI methods, and none are based on bootstrapping.
-- New sffect sizes for contingency tables (`phi()` and `cramers_v()`).
-- `chisq_to_phi/cramers_v()` functions now support CIs (via the ncp method), and return a data frame.
+- New effect sizes for contingency tables (`phi()` and `cramers_v()`).
+- `chisq_to_phi()` / `cramers_v()` functions now support CIs (via the ncp method), and return a data frame.
 - `F_to_eta2()` family of functions now support CIs (via the ncp method), and return a data frame.
 - `t_to_d()` and `t_to_r()` now support CIs (via the ncp method), and return a data frame.
 - `standardize()` for model-objects has a default-method, which usually accepts all models. Exception for model-objects that do not work will be added if missing.
 - `standardize.data.frame()` gets `append` and `suffix` arguments, to add (instead of replace) standardized variables to the returned data frame.
+- `eta_squared()`, `omega_squared()` and `epsilon_squared()` now works for output from `parameters::model_parameters()`.
 
 ## Bug fixes