+++ NEWS.md
@@ -11,7 +11,9 @@
 - `t_to_d()` and `t_to_r()` now support CIs (via the ncp method), and return a data frame.
 - `standardize()` for model-objects has a default-method, which usually accepts all models. Exception for model-objects that do not work will be added if missing.
 - `standardize.data.frame()` gets `append` and `suffix` arguments, to add (instead of replace) standardized variables to the returned data frame.
-- `eta_squared()`, `omega_squared()` and `epsilon_squared()` now works for output from `parameters::model_parameters()`.
+- `eta_squared()`, `omega_squared()` and `epsilon_squared()` now works
+  - output from `parameters::model_parameters()`.
+  - `mlm` models.
 
 ## Bug fixes